Create Algorithm to count the number of triplets
The face value may range from 2-14
 The suit has four possible values. 
You wish to know the total number of unique triplets (a set of three cards sharing the same face value and same suit) that are inside the input hand. 
Order does not matter inside the triplet, triplet {C1, C2, C3} is the same as {C2,C1,C3}. 
This algorithm generates this total.